Add base entity to workday
